# frontend/openapi.json
{
  "components": {
    "schemas": {
      "EpisodeRead": {
        "description": "Public representation of a podcast episode.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "episode_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Number"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "podcast_id": {
            "title": "Podcast Id",
            "type": "integer"
          },
          "podcast_public_id": {
            "description": "The opaque public identifier of the owning podcast.",
            "readOnly": true,
            "title": "Podcast Public Id",
            "type": "string"
          },
          "public_id": {
            "description": "The opaque, prefixed public identifier for this episode.",
            "readOnly": true,
            "title": "Public Id",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "podcast_id",
          "title",
          "episode_number",
          "published_at",
          "created_at",
          "public_id",
          "podcast_public_id"
        ],
        "title": "EpisodeRead",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MediaRead": {
        "description": "Public representation of a canonical media item.",
        "properties": {
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "public_id": {
            "description": "The opaque, prefixed public identifier for this media item.",
            "readOnly": true,
            "title": "Public Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/MediaType"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "id",
          "type",
          "title",
          "author",
          "year",
          "description",
          "cover_url",
          "created_at",
          "public_id"
        ],
        "title": "MediaRead",
        "type": "object"
      },
      "MediaType": {
        "description": "The kind of media a catalog item represents.",
        "enum": [
          "book",
          "movie",
          "documentary",
          "tv_show",
          "study",
          "podcast",
          "article",
          "person",
          "place"
        ],
        "title": "MediaType",
        "type": "string"
      },
      "MentionRead": {
        "description": "Public representation of a media mention within an episode.",
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "episode_id": {
            "title": "Episode Id",
            "type": "integer"
          },
          "episode_public_id": {
            "description": "The opaque public identifier of the referenced episode.",
            "readOnly": true,
            "title": "Episode Public Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "media_id": {
            "title": "Media Id",
            "type": "integer"
          },
          "media_public_id": {
            "description": "The opaque public identifier of the referenced media item.",
            "readOnly": true,
            "title": "Media Public Id",
            "type": "string"
          },
          "public_id": {
            "description": "The opaque, prefixed public identifier for this mention.",
            "readOnly": true,
            "title": "Public Id",
            "type": "string"
          },
          "timestamp_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp Seconds"
          }
        },
        "required": [
          "id",
          "episode_id",
          "media_id",
          "timestamp_seconds",
          "context",
          "confidence",
          "created_at",
          "public_id",
          "episode_public_id",
          "media_public_id"
        ],
        "title": "MentionRead",
        "type": "object"
      },
      "Page_EpisodeRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EpisodeRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[EpisodeRead]",
        "type": "object"
      },
      "Page_MediaRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MediaRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[MediaRead]",
        "type": "object"
      },
      "Page_MentionRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MentionRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[MentionRead]",
        "type": "object"
      },
      "Page_PodcastRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PodcastRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[PodcastRead]",
        "type": "object"
      },
      "PodcastRead": {
        "description": "Public representation of a podcast source.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "public_id": {
            "description": "The opaque, prefixed public identifier for this podcast.",
            "readOnly": true,
            "title": "Public Id",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "created_at",
          "public_id"
        ],
        "title": "PodcastRead",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "podex",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v2/episodes": {
      "get": {
        "description": "List episodes, optionally filtered by podcast, paginated.",
        "operationId": "list_episodes_api_v2_episodes_get",
        "parameters": [
          {
            "in": "query",
            "name": "podcast_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Podcast Id"
            }
          },
          {
            "description": "Maximum number of items to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Maximum number of items to return.",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Number of items to skip.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Number of items to skip.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_EpisodeRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Episodes",
        "tags": [
          "v2",
          "episodes"
        ]
      }
    },
    "/api/v2/episodes/{episode_id}": {
      "get": {
        "description": "Return a single episode by id.",
        "operationId": "get_episode_api_v2_episodes__episode_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Episode",
        "tags": [
          "v2",
          "episodes"
        ]
      }
    },
    "/api/v2/episodes/{episode_id}/mentions": {
      "get": {
        "description": "List media mentions within an episode, ordered by timestamp, paginated.",
        "operationId": "list_episode_mentions_api_v2_episodes__episode_id__mentions_get",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "integer"
            }
          },
          {
            "description": "Maximum number of items to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Maximum number of items to return.",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Number of items to skip.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Number of items to skip.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MentionRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Episode Mentions",
        "tags": [
          "v2",
          "episodes"
        ]
      }
    },
    "/api/v2/media": {
      "get": {
        "description": "List media items, optionally filtered by type, paginated.",
        "operationId": "list_media_api_v2_media_get",
        "parameters": [
          {
            "in": "query",
            "name": "media_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MediaType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Media Type"
            }
          },
          {
            "description": "Maximum number of items to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Maximum number of items to return.",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Number of items to skip.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Number of items to skip.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MediaRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Media",
        "tags": [
          "v2",
          "media"
        ]
      }
    },
    "/api/v2/media/{media_id}": {
      "get": {
        "description": "Return a single media item by id.",
        "operationId": "get_media_api_v2_media__media_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "media_id",
            "required": true,
            "schema": {
              "title": "Media Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Media",
        "tags": [
          "v2",
          "media"
        ]
      }
    },
    "/api/v2/media/{media_id}/mentions": {
      "get": {
        "description": "List episode mentions of a media item, paginated.",
        "operationId": "list_media_mentions_api_v2_media__media_id__mentions_get",
        "parameters": [
          {
            "in": "path",
            "name": "media_id",
            "required": true,
            "schema": {
              "title": "Media Id",
              "type": "integer"
            }
          },
          {
            "description": "Maximum number of items to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Maximum number of items to return.",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Number of items to skip.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Number of items to skip.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MentionRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Media Mentions",
        "tags": [
          "v2",
          "media"
        ]
      }
    },
    "/api/v2/podcasts": {
      "get": {
        "description": "List podcast sources ordered by name, paginated by ``limit``/``offset``.",
        "operationId": "list_podcasts_api_v2_podcasts_get",
        "parameters": [
          {
            "description": "Maximum number of items to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Maximum number of items to return.",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Number of items to skip.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Number of items to skip.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PodcastRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Podcasts",
        "tags": [
          "v2",
          "podcasts"
        ]
      }
    },
    "/api/v2/podcasts/{podcast_id}": {
      "get": {
        "description": "Return a single podcast source by id.",
        "operationId": "get_podcast_api_v2_podcasts__podcast_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "podcast_id",
            "required": true,
            "schema": {
              "title": "Podcast Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PodcastRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Podcast",
        "tags": [
          "v2",
          "podcasts"
        ]
      }
    },
    "/api/v2/status": {
      "get": {
        "description": "Report that the v2 API surface is reachable.",
        "operationId": "read_status_api_v2_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Read Status Api V2 Status Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Read Status",
        "tags": [
          "v2"
        ]
      }
    },
    "/health": {
      "get": {
        "description": "Liveness probe.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    }
  }
}